+++ builder/action.py
@@ -85,7 +85,7 @@ class Action(BaseContainer):
                     tmp.append(Then())
                 elif v.startswith('#'):
                     ## NOTE: current info only
-                    tmp.append(MetaData(MetaType.INFO, v.replace('#', '＃', 1)))
+                    tmp.append(MetaData(MetaType.INFO, note=v.replace('#', '＃', 1)))
                 else:
                     tmp.append(v)
             elif isinstance(v, AllSubjects):